sandbox: UInt8 is too small to hold a TCP or UDP port
TCP and UDP port numbers are unsigned 16-bit integers, but
SandboxNetworkPermission uses UInt8s so trying to specify anything
but a small subset of low port numbers causes an error:

    permissions: [
        .allowNetworkConnections(
            scope: .all(ports: [443, 8080]),
            reason: "This command accesses the network"
        )
    ]

leads to

    error: integer literal '443' overflows when stored into 'Array<UInt8>.ArrayLiteralElement' (aka 'UInt8')
    error: integer literal '8080' overflows when stored into 'Array<UInt8>.ArrayLiteralElement' (aka 'UInt8')

This commit updates the type definitions and adds some higher-numbered
ports to the .allowNetworkConnections test cases.

+++ Sources/Basics/Sandbox.swift
@@ -15,8 +15,8 @@ import func TSCBasic.determineTempDirectory
 
 public enum SandboxNetworkPermission: Equatable {
     case none
-    case local(ports: [UInt8])
-    case all(ports: [UInt8])
+    case local(ports: [Int])
+    case all(ports: [Int])
     case docker
     case unixDomainSocket
 
@@ -28,7 +28,7 @@ public enum SandboxNetworkPermission: Equatable {
         }
     }
 
-    fileprivate var ports: [UInt8] {
+    fileprivate var ports: [Int] {
         switch self {
         case .all(let ports): return ports
         case .local(let ports): return ports
